refactor(lexical-graph): split chunk store factories into per-backend files
InGraphChunkStoreFactory and S3ChunkStoreFactory each get their own
module under storage/chunk/, mirroring the per-backend split of the
stores themselves and the vector layer's s3_vector_index_factory.
chunk_store_factory.py keeps the registry and ChunkStoreFactory.
